Add SmartThingSchedule and different fixes (#39)
* Use time points instead of intervals

* Add SmartThinkShedule

* Change ParseString and add ThingsSchedule

* Update schedule

* Change DataPacket

* Fix SmartThingsSchedule and uncomment DataPacket tests

* Fix time and increase performance

* Update schedule

* Add some sh...

* TAKE IT VLAD

* Update

* Add Object Tests

* Update tests

* Add some tests

* ExternalAction and ThingsSchedule tests are added

* Fix linux fails

* ExternalAction and ThingsSchedule tests are added

* Fix pointers

+++ Dll/Road/Car.cpp
@@ -24,7 +24,7 @@ void TCar::Update()
 	TObjectOfObservation::Update();
 	auto&& val = this->properties["WayIndex"]->GetValues();
 	std::default_random_engine generator;
-	generator.seed(std::chrono::high_resolution_clock::now().time_since_epoch().count());
+	generator.seed(static_cast<unsigned int>(std::chrono::high_resolution_clock::now().time_since_epoch().count()));
 	std::poisson_distribution<int> poisson;
 	val["WayIndex"] = poisson(generator) % static_cast<int>(this->properties["WayIndexCount"]->GetValues()["WayIndexCount"]);
 	this->properties["WayIndex"]->SetValues(val);
